insight_bot/requirements.txt
py-cord
google-genai
pydub
python-dotenv
PyNaCl
